Picking up objects
Should be able to pick up the object even though it's pretty buggy at the moment

+++ Assets/Fps Controller/Fps Scripts/FpsMovment.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class FpsMovment : MonoBehaviour
 {
     Rigidbody rb;
-    [SerializeField] int MoveSpeed;
+    [SerializeField] float MoveSpeed;
 
 
     // Start is called before the first frame update
